Specify bypass URL port in CrConfig only if explicitly set
This fix makes Traffic Ops specify the 'port' field in the 'bypass' section of
CrConfig only if explicitly set in the DS 'Bypass FQDN' field (By using the format 'fqdn:port').

Before the fix, if a port was not explicitly set in the DS 'Bypass FQDN' field,
Ops specified a value of 80 to the 'port' field in CrConfig, making HTTPS requests that
go to bypass being redirected to https://<bypass-fqdn>:80/... , which is
certainly not right.

Resolves: #1493 TR redirects to Bypass FQDN with port 80 for HTTPS DSes